(PUP-9212) Allow for image entry point CMDs
Prior to this commit, a user could choose between running from their own
custom entrypoint or running from the default ssh CMD. This made it
probelmatic for users wanting to use the default entrypoint for on image
while also having ssh running in the container.

This change allows a user to set 'use_image_entry_point' for a host;
this will use the image's own default CMD and load ssh after the image
has started, so that Beaker's DSL can still communicate with the
container.