(PUP-9212) Allow for building containers with context
Before this commit, beaker-docker would copy the file contents of a
supplied dockerfile and use that to generate a new image. This had the
drawback of not being able to use much of the Dockerfile DSL, such as
COPY or VOLUME, because they were relative paths that were not copied
along with the Dockerfile itself.

This commit changes it so that a dockerfile is copied along with the
context of the test execution directory, so that those Dockerfile DSL
methods are usable with a Dockerfile used by beaker.