(BKR-1551) Don't expect to receive a logger instance from options
Before subcommands were added in Beaker 4, storing objects like loggers
in the host hash worked fine; The docker hypervisor expected to receive
a logger instance this way during initialization.

This no longer works now that stages of a beaker run are split into
separate commands that persist options in subcommand_options.yaml, since
a logger instance read from the YAML file can't and shouldn't be
reconstructed. This change preserves the option to specify a custom
logger in the constructor, but falls back to a new logger otherwise.