(BKR-1551) Don't persist container objects to yaml
Previously, docker container objects were stored for easy access in
host['docker_container'].

In Beaker 4, the host hash is persisted to subcommand_options.yaml
between runs of beaker subcommands. This causes the container object to
be written as a string, which becomes unusable when loaded from the YAML
file during subsequent subcommand runs.

For compatibility with Beaker 4:
- Record the container ID in the host hash as 'docker_container_id'
  (instead of recording the entire container object as
  host['docker_container']) while provisioning.
- Refactor the find_container method to locate an existing container by
  either a name or a an ID, instead of only by name.