(BKR-1551) Don't persist image objects to yaml
Previously, docker image objects were stored for easy access in
host['docker_image'].

In Beaker 4, the host hash is persisted to subcommand_options.yaml
between runs of beaker subcommands. This causes the image object to
be written as a string, which becomes unusable when loaded from the YAML
file during subsequent subcommand runs.

For compatibility with Beaker 4:
- Record the image ID in the host hash as 'docker_image_id' (instead of
  recording the entire image object as host['docker_image']) while
  provisioning.
- Add a find_image method to locate an existing image by its ID, instead
  of relying on a saved image object