(maint) A number of fixes for rerunning tests on docker containers (#18)
* (maint) Only provision container if it can't be found

Previously, this would provision the container if it wasn't found or if
provisioning was explicitly enabled, which is the default option. It
feels like you shouldn't need to explicitly disable provisioning to
reuse running containers.

* (maint) Add support for setting Env for containers

* (maint) Set `use_existing_container` in the host object

This will allow us to determine elsewhere if we're using an existing
container and can skip provisioning steps.

* (maint) Add support for `docker_preserve_container`

This is equivalent to preserving the host for non-containerized runs. If
this option is set, the container will not be destroyed on test
completion and the image the container is based on will not be removed.

* (maint) `find_container` needs to explicitly return nil when not found