Cleanup docs and gemspec
Emails were removed from the gemspec because they are no longer valid
addresses. The author names were preserved to represent those who have
put lots of work into this.

Other items were updated to reflect that this gem was recently dontated
to Vox Pupuli and to account for recent PRs.